// src/yowyob-billing/models/LigneBonAchatRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type LigneBonAchatRequest = {
    productId?: string;
    productCode?: string;
    productName?: string;
    uom?: string;
    orderedQuantity?: number;
    unitPrice?: number;
    taxable?: boolean;
    vatAmount?: number;
    totalAmount?: number;
};