// src/yowyob-billing/models/LineBonReception.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type LineBonReception = {
    productId?: string;
    description?: string;
    uom?: string;
    orderedQuantity?: number;
    receivedQuantity?: number;
    acceptedQuantity?: number;
    rejectedQuantity?: number;
    shortQuantity?: number;
    damagedQuantity?: number;
    excessQuantity?: number;
    rate?: number;
    discount?: number;
    lineAmount?: number;
};